refactor: improve directory browser organization
- Split DirectoryBrowser into smaller components (PathDisplay, ErrorDisplay, NavigationButtons, DialogFooter)
- Created shared types and hooks
- Improved component organization with proper folder structure
- Added index file for exports
